user registration functionality
It was added the enabled field to the user model. The password
constraint was developed, but It needs to be improved. The
UserValidator was added, but  It also needs to be improved. It
was changed the register user view and the user routting. The
functionality to register the user is already running.